requirements.txt
dash==2.14.1
plotly==5.17.0
sqlalchemy==2.0.23
pandas==2.1.1
numpy==1.24.3
python-dateutil==2.8.2
